refactor(lifecycle): extract keymap ownership from accessors
accessors.lua held two unrelated jobs: reading and writing session
fields, and managing the lifetime of the session's keymap registry.
The second group never touched a session field — every function looked
the session up by tabpage and delegated straight to codediff.keymap.

Move those 13 functions to lifecycle/keymaps.lua, so the boundary reads:

  keymap/*             key engine, unaware of sessions
  lifecycle/keymaps    tabpage -> registry, and binds registry lifetime
                       to the session (suspend on TabLeave, dispose on
                       cleanup)
  lifecycle/accessors  session field access only

Pure move; no logic changed. The lifecycle facade absorbs it, so the 77
call sites going through require("codediff.ui.lifecycle") are
untouched. Three in-package call sites that bypassed the facade were
repointed.

accessors.lua 834 -> 615 lines.
