+++ lua/codediff/ui/lifecycle/cleanup.lua
@@ -2,6 +2,7 @@
 local M = {}
 
 local accessors = require("codediff.ui.lifecycle.accessors")
+local keymaps = require("codediff.ui.lifecycle.keymaps")
 local session = require("codediff.ui.lifecycle.session")
 local state = require("codediff.ui.lifecycle.state")
 local welcome_window = require("codediff.ui.view.welcome_window")
@@ -48,7 +49,7 @@ local function cleanup_diff(tabpage)
   state.restore_buffer_state(diff.modified_bufnr, diff.modified_state)
 
   -- Hand every mapped key back to whatever owned it before codediff
-  accessors.dispose_keymaps(tabpage)
+  keymaps.dispose_keymaps(tabpage)
 
   -- Call explorer's cleanup function to stop file watchers
   if diff.explorer and diff.explorer._cleanup_auto_refresh then
